==> code/Step3_Within_Area_Estimates_for_All_Chicago.R <==
rm(list = ls())

#Load Necessary Packages
library(plyr)
library(doBy)
library(reshape)
library(glmmML)


setwd("C:/Users/Zach/Documents/UrbanCCD/Streetlights")

#Open Alley Lights and Crime
Alley.Lights<- read.csv(file="alley.csv", head=TRUE)

#Open Street Lights One Out and Crime
Street.Lights.OneOut<- read.csv(file="street-one.csv", head=TRUE)

#Open Street Lights All Out and Crime
Street.Lights.AllOut<- read.csv(file="street-all.csv", head=TRUE)

#Change Community Areas to Numeric
Alley.Lights$community_area <- as.character(Alley.Lights$community_area)
Street.Lights.OneOut$community_area <- as.character(Street.Lights.OneOut$community_area)
Street.Lights.AllOut$community_area <- as.character(Street.Lights.AllOut$community_area)

Alley.Lights$community_area[which(nchar(Alley.Lights$community_area)==4)] <- as.numeric(substr(Alley.Lights$community_area[which(nchar(Alley.Lights$community_area)==4)],3,3))
Alley.Lights$community_area[which(nchar(Alley.Lights$community_area)==5)] <- as.numeric(substr(Alley.Lights$community_area[which(nchar(Alley.Lights$community_area)==5)],3,4))
Street.Lights.OneOut$community_area[which(nchar(Street.Lights.OneOut$community_area)==4)] <- as.numeric(substr(Street.Lights.OneOut$community_area[which(nchar(Street.Lights.OneOut$community_area)==4)],3,3))
Street.Lights.OneOut$community_area[which(nchar(Street.Lights.OneOut$community_area)==5)] <- as.numeric(substr(Street.Lights.OneOut$community_area[which(nchar(Street.Lights.OneOut$community_area)==5)],3,4))
Street.Lights.AllOut$community_area[which(nchar(Street.Lights.AllOut$community_area)==4)] <- as.numeric(substr(Street.Lights.AllOut$community_area[which(nchar(Street.Lights.AllOut$community_area)==4)],3,3))
Street.Lights.AllOut$community_area[which(nchar(Street.Lights.AllOut$community_area)==5)] <- as.numeric(substr(Street.Lights.AllOut$community_area[which(nchar(Street.Lights.AllOut$community_area)==5)],3,4))

# Remove Duplicates
Alley.Lights         <-         Alley.Lights[!duplicated(Alley.Lights$Service.Request.No),]
Street.Lights.OneOut <- Street.Lights.OneOut[!duplicated(Street.Lights.OneOut$Service.Request.No),]
Street.Lights.AllOut <- Street.Lights.AllOut[!duplicated(Street.Lights.AllOut$Service.Request.No),]


# Remove Community Area 0
Alley.Lights         <-         Alley.Lights[Alley.Lights$community_area        !=0,]
Street.Lights.OneOut <- Street.Lights.OneOut[Street.Lights.OneOut$community_area!=0,]
Street.Lights.AllOut <- Street.Lights.AllOut[Street.Lights.AllOut$community_area!=0,]


#Create Summary Tables
Summary.Table.Alley  <- data.frame(matrix(ncol = 4, nrow = 8), row.names = c("All Crimes (No Deceptive Practice)", 
                                                                              "Theft", "Narcotics", "Battery", "Criminal Damage", "Motor Vehicle Theft", 
                                                                               "Robbery", "Assault"))
Summary.Table.OneOut <- data.frame(matrix(ncol = 4, nrow = 11), row.names = c("All Crimes (No Deceptive Practice)", 
                                                                              "Theft", "Narcotics", "Battery", "Criminal Damage", "Motor Vehicle Theft", 
                                                                              "Robbery", "Assault", "Burglary", "Homicide", "Deceptive Practice"))
Summary.Table.AllOut <- data.frame(matrix(ncol = 4, nrow = 11), row.names = c("All Crimes (No Deceptive Practice)", 
                                                                              "Theft", "Narcotics", "Battery", "Criminal Damage", "Motor Vehicle Theft", 
                                                                              "Robbery", "Assault", "Burglary", "Homicide", "Deceptive Practice"))
Summary.Table.Alley  <- rename(Summary.Table.Alley,  c("X1"="Before", "X2"="During", "X3"="After", "X4"="AvgBeforeAfter"))
Summary.Table.OneOut <- rename(Summary.Table.OneOut, c("X1"="Before", "X2"="During", "X3"="After", "X4"="AvgBeforeAfter"))
Summary.Table.AllOut <- rename(Summary.Table.AllOut, c("X1"="Before", "X2"="During", "X3"="After", "X4"="AvgBeforeAfter"))


Summary.Table.Alley[1,1]  <- 30*sum(Alley.Lights$Crimes.All.Before)       /(30*nrow(Alley.Lights))
Summary.Table.Alley[2,1]  <- 30*sum(Alley.Lights$Theft.Before)            /(30*nrow(Alley.Lights))
Summary.Table.Alley[3,1]  <- 30*sum(Alley.Lights$Narcotics.Before)        /(30*nrow(Alley.Lights))
Summary.Table.Alley[4,1]  <- 30*sum(Alley.Lights$Battery.Before)          /(30*nrow(Alley.Lights))
Summary.Table.Alley[5,1]  <- 30*sum(Alley.Lights$CriminalDamage.Before)   /(30*nrow(Alley.Lights))
Summary.Table.Alley[6,1]  <- 30*sum(Alley.Lights$MotorVehicleTheft.Before)/(30*nrow(Alley.Lights))
Summary.Table.Alley[7,1]  <- 30*sum(Alley.Lights$Robbery.Before)          /(30*nrow(Alley.Lights))
Summary.Table.Alley[8,1]  <- 30*sum(Alley.Lights$Assault.Before)          /(30*nrow(Alley.Lights))
Summary.Table.Alley[1,2]  <- 30*sum(Alley.Lights$Crimes.All.During)       /sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[2,2]  <- 30*sum(Alley.Lights$Theft.During)           /sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[3,2]  <- 30*sum(Alley.Lights$Narcotics.During)        /sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[4,2]  <- 30*sum(Alley.Lights$Battery.During)          /sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[5,2]  <- 30*sum(Alley.Lights$CriminalDamage.During)   /sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[6,2]  <- 30*sum(Alley.Lights$MotorVehicleTheft.During)/sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[7,2]  <- 30*sum(Alley.Lights$Robbery.During)           /sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[8,2]  <- 30*sum(Alley.Lights$Assault.During)           /sum(Alley.Lights$OutageDuration)
Summary.Table.Alley[1,3]  <- 30*sum(Alley.Lights$Crimes.All.After)         /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[2,3]  <- 30*sum(Alley.Lights$Theft.After)              /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[3,3]  <- 30*sum(Alley.Lights$Narcotics.After)          /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[4,3]  <- 30*sum(Alley.Lights$Battery.After)            /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[5,3]  <- 30*sum(Alley.Lights$CriminalDamage.After)     /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[6,3]  <- 30*sum(Alley.Lights$MotorVehicleTheft.After)  /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[7,3]  <- 30*sum(Alley.Lights$Robbery.After)            /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[8,3]  <- 30*sum(Alley.Lights$Assault.After)            /sum(Alley.Lights$After.Period.Duration)
Summary.Table.Alley[1,4]  <- 30*(sum(Alley.Lights$Crimes.All.Before)       +sum(Alley.Lights$Crimes.All.After))       /(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))
Summary.Table.Alley[2,4]  <- 30*(sum(Alley.Lights$Theft.Before)            +sum(Alley.Lights$Theft.After))            /(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))
Summary.Table.Alley[3,4]  <- 30*(sum(Alley.Lights$Narcotics.Before)        +sum(Alley.Lights$Narcotics.After))        /(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))
Summary.Table.Alley[4,4]  <- 30*(sum(Alley.Lights$Battery.Before)          +sum(Alley.Lights$Battery.After))          /(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))
Summary.Table.Alley[5,4]  <- 30*(sum(Alley.Lights$CriminalDamage.Before)   +sum(Alley.Lights$CriminalDamage.After))   /(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))
Summary.Table.Alley[6,4]  <- 30*(sum(Alley.Lights$MotorVehicleTheft.Before)+sum(Alley.Lights$MotorVehicleTheft.After))/(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))
Summary.Table.Alley[7,4]  <- 30*(sum(Alley.Lights$Robbery.Before)          +sum(Alley.Lights$Robbery.After))          /(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))
Summary.Table.Alley[8,4]  <- 30*(sum(Alley.Lights$Assault.Before)          +sum(Alley.Lights$Assault.After))          /(30*nrow(Alley.Lights)+sum(Alley.Lights$After.Period.Duration))

Summary.Table.OneOut[1,1]  <- 30*sum(Street.Lights.OneOut$Crimes.All.Before)       /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[2,1]  <- 30*sum(Street.Lights.OneOut$Theft.Before)           /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[3,1]  <- 30*sum(Street.Lights.OneOut$Narcotics.Before)        /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[4,1]  <- 30*sum(Street.Lights.OneOut$Battery.Before)          /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[5,1]  <- 30*sum(Street.Lights.OneOut$CriminalDamage.Before)   /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[6,1]  <- 30*sum(Street.Lights.OneOut$MotorVehicleTheft.Before)/(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[7,1]  <- 30*sum(Street.Lights.OneOut$Robbery.Before)          /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[8,1]  <- 30*sum(Street.Lights.OneOut$Assault.Before)           /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[9,1]  <- 30*sum(Street.Lights.OneOut$Burglary.Before)          /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[10,1] <- 30*sum(Street.Lights.OneOut$Homicide.Before)         /(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[11,1] <- 30*sum(Street.Lights.OneOut$DeceptivePractice.Before)/(30*nrow(Street.Lights.OneOut))
Summary.Table.OneOut[1,2]  <- 30*sum(Street.Lights.OneOut$Crimes.All.During)       /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[2,2]  <- 30*sum(Street.Lights.OneOut$Theft.During)           /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[3,2]  <- 30*sum(Street.Lights.OneOut$Narcotics.During)        /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[4,2]  <- 30*sum(Street.Lights.OneOut$Battery.During)          /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[5,2]  <- 30*sum(Street.Lights.OneOut$CriminalDamage.During)   /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[6,2]  <- 30*sum(Street.Lights.OneOut$MotorVehicleTheft.During)/sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[7,2]  <- 30*sum(Street.Lights.OneOut$Robbery.During)          /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[8,2]  <- 30*sum(Street.Lights.OneOut$Assault.During)           /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[9,2]  <- 30*sum(Street.Lights.OneOut$Burglary.During)          /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[10,2] <- 30*sum(Street.Lights.OneOut$Homicide.During)         /sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[11,2] <- 30*sum(Street.Lights.OneOut$DeceptivePractice.During)/sum(Street.Lights.OneOut$OutageDuration)
Summary.Table.OneOut[1,3]  <- 30*sum(Street.Lights.OneOut$Crimes.All.After)        /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[2,3]  <- 30*sum(Street.Lights.OneOut$Theft.After)            /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[3,3]  <- 30*sum(Street.Lights.OneOut$Narcotics.After)         /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[4,3]  <- 30*sum(Street.Lights.OneOut$Battery.After)           /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[5,3]  <- 30*sum(Street.Lights.OneOut$CriminalDamage.After)    /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[6,3]  <- 30*sum(Street.Lights.OneOut$MotorVehicleTheft.After) /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[7,3]  <- 30*sum(Street.Lights.OneOut$Robbery.After)           /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[8,3]  <- 30*sum(Street.Lights.OneOut$Assault.After)           /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[9,3]  <- 30*sum(Street.Lights.OneOut$Burglary.After)          /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[10,3] <- 30*sum(Street.Lights.OneOut$Homicide.After)          /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[11,3] <- 30*sum(Street.Lights.OneOut$DeceptivePractice.After) /sum(Street.Lights.OneOut$After.Period.Duration)
Summary.Table.OneOut[1,4]  <- 30*(sum(Street.Lights.OneOut$Crimes.All.Before)       +sum(Street.Lights.OneOut$Crimes.All.After))       /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[2,4]  <- 30*(sum(Street.Lights.OneOut$Theft.Before)           +sum(Street.Lights.OneOut$Theft.After))           /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[3,4]  <- 30*(sum(Street.Lights.OneOut$Narcotics.Before)        +sum(Street.Lights.OneOut$Narcotics.After))        /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[4,4]  <- 30*(sum(Street.Lights.OneOut$Battery.Before)          +sum(Street.Lights.OneOut$Battery.After))          /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[5,4]  <- 30*(sum(Street.Lights.OneOut$CriminalDamage.Before)   +sum(Street.Lights.OneOut$CriminalDamage.After))   /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[6,4]  <- 30*(sum(Street.Lights.OneOut$MotorVehicleTheft.Before)+sum(Street.Lights.OneOut$MotorVehicleTheft.After))/(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[7,4]  <- 30*(sum(Street.Lights.OneOut$Robbery.Before)          +sum(Street.Lights.OneOut$Robbery.After))          /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[8,4]  <- 30*(sum(Street.Lights.OneOut$Assault.Before)           +sum(Street.Lights.OneOut$Assault.After))          /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[9,4]  <- 30*(sum(Street.Lights.OneOut$Burglary.Before)          +sum(Street.Lights.OneOut$Burglary.After))         /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[10,4] <- 30*(sum(Street.Lights.OneOut$Homicide.Before)         +sum(Street.Lights.OneOut$Homicide.After))         /(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))
Summary.Table.OneOut[11,4] <- 30*(sum(Street.Lights.OneOut$DeceptivePractice.Before)+sum(Street.Lights.OneOut$DeceptivePractice.After))/(30*nrow(Street.Lights.OneOut)+sum(Street.Lights.OneOut$After.Period.Duration))


Summary.Table.AllOut[1,1]  <- 30*sum(Street.Lights.AllOut$Crimes.All.Before)       /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[2,1]  <- 30*sum(Street.Lights.AllOut$Theft.Before)           /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[3,1]  <- 30*sum(Street.Lights.AllOut$Narcotics.Before)        /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[4,1]  <- 30*sum(Street.Lights.AllOut$Battery.Before)          /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[5,1]  <- 30*sum(Street.Lights.AllOut$CriminalDamage.Before)   /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[6,1]  <- 30*sum(Street.Lights.AllOut$MotorVehicleTheft.Before)/(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[7,1]  <- 30*sum(Street.Lights.AllOut$Robbery.Before)          /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[8,1]  <- 30*sum(Street.Lights.AllOut$Assault.Before)          /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[9,1]  <- 30*sum(Street.Lights.AllOut$Burglary.Before)         /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[10,1] <- 30*sum(Street.Lights.AllOut$Homicide.Before)         /(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[11,1] <- 30*sum(Street.Lights.AllOut$DeceptivePractice.Before)/(30*nrow(Street.Lights.AllOut))
Summary.Table.AllOut[1,2]  <- 30*sum(Street.Lights.AllOut$Crimes.All.During)       /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[2,2]  <- 30*sum(Street.Lights.AllOut$Theft.During)           /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[3,2]  <- 30*sum(Street.Lights.AllOut$Narcotics.During)        /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[4,2]  <- 30*sum(Street.Lights.AllOut$Battery.During)          /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[5,2]  <- 30*sum(Street.Lights.AllOut$CriminalDamage.During)   /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[6,2]  <- 30*sum(Street.Lights.AllOut$MotorVehicleTheft.During)/sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[7,2]  <- 30*sum(Street.Lights.AllOut$Robbery.During)          /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[8,2]  <- 30*sum(Street.Lights.AllOut$Assault.During)          /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[9,2]  <- 30*sum(Street.Lights.AllOut$Burglary.During)         /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[10,2] <- 30*sum(Street.Lights.AllOut$Homicide.During)         /sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[11,2] <- 30*sum(Street.Lights.AllOut$DeceptivePractice.During)/sum(Street.Lights.AllOut$OutageDuration)
Summary.Table.AllOut[1,3]  <- 30*sum(Street.Lights.AllOut$Crimes.All.After)        /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[2,3]  <- 30*sum(Street.Lights.AllOut$Theft.After)            /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[3,3]  <- 30*sum(Street.Lights.AllOut$Narcotics.After)         /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[4,3]  <- 30*sum(Street.Lights.AllOut$Battery.After)           /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[5,3]  <- 30*sum(Street.Lights.AllOut$CriminalDamage.After)    /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[6,3]  <- 30*sum(Street.Lights.AllOut$MotorVehicleTheft.After) /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[7,3]  <- 30*sum(Street.Lights.AllOut$Robbery.After)           /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[8,3]  <- 30*sum(Street.Lights.AllOut$Assault.After)           /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[9,3]  <- 30*sum(Street.Lights.AllOut$Burglary.After)          /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[10,3] <- 30*sum(Street.Lights.AllOut$Homicide.After)          /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[11,3] <- 30*sum(Street.Lights.AllOut$DeceptivePractice.After) /sum(Street.Lights.AllOut$After.Period.Duration)
Summary.Table.AllOut[1,4]  <- 30*(sum(Street.Lights.AllOut$Crimes.All.Before)       +sum(Street.Lights.AllOut$Crimes.All.After))       /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[2,4]  <- 30*(sum(Street.Lights.AllOut$Theft.Before)            +sum(Street.Lights.AllOut$Theft.After))            /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[3,4]  <- 30*(sum(Street.Lights.AllOut$Narcotics.Before)        +sum(Street.Lights.AllOut$Narcotics.After))        /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[4,4]  <- 30*(sum(Street.Lights.AllOut$Battery.Before)          +sum(Street.Lights.AllOut$Battery.After))          /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[5,4]  <- 30*(sum(Street.Lights.AllOut$CriminalDamage.Before)   +sum(Street.Lights.AllOut$CriminalDamage.After))   /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[6,4]  <- 30*(sum(Street.Lights.AllOut$MotorVehicleTheft.Before)+sum(Street.Lights.AllOut$MotorVehicleTheft.After))/(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[7,4]  <- 30*(sum(Street.Lights.AllOut$Robbery.Before)          +sum(Street.Lights.AllOut$Robbery.After))          /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[8,4]  <- 30*(sum(Street.Lights.AllOut$Assault.Before)          +sum(Street.Lights.AllOut$Assault.After))          /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[9,4]  <- 30*(sum(Street.Lights.AllOut$Burglary.Before)         +sum(Street.Lights.AllOut$Burglary.After))         /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[10,4] <- 30*(sum(Street.Lights.AllOut$Homicide.Before)         +sum(Street.Lights.AllOut$Homicide.After))         /(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))
Summary.Table.AllOut[11,4] <- 30*(sum(Street.Lights.AllOut$DeceptivePractice.Before)+sum(Street.Lights.AllOut$DeceptivePractice.After))/(30*nrow(Street.Lights.AllOut)+sum(Street.Lights.AllOut$After.Period.Duration))

Summary.Table.Alley$AbsDiff  <- Summary.Table.Alley$During  - Summary.Table.Alley$AvgBeforeAfter
Summary.Table.OneOut$AbsDiff <- Summary.Table.OneOut$During - Summary.Table.OneOut$AvgBeforeAfter
Summary.Table.AllOut$AbsDiff <- Summary.Table.AllOut$During - Summary.Table.AllOut$AvgBeforeAfter




#Create Datasets Formatted for Poisson Modeling and Estimation of SE's
Alley.Lights.Pois <- Alley.Lights[rep(seq(nrow(Alley.Lights)),3),]
Alley.Lights.Pois <- Alley.Lights.Pois[order(Alley.Lights.Pois$Service.Request.No),]
Alley.Lights.Pois$Time              <- rep(c("BEFORE", "DURING", "AFTER"), nrow(Alley.Lights))
Alley.Lights.Pois$AllCrimes         <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$Theft             <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$Narcotics         <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$Battery           <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$CriminalDamage    <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$MotorVehicleTheft <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$Robbery           <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$Assault           <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$Duration          <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$AllCrimes[which(Alley.Lights.Pois$Time=="BEFORE")]         <- Alley.Lights.Pois$Crimes.All.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$Theft[which(Alley.Lights.Pois$Time=="BEFORE")]            <- Alley.Lights.Pois$Theft.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$Narcotics[which(Alley.Lights.Pois$Time=="BEFORE")]         <- Alley.Lights.Pois$Narcotics.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$Battery[which(Alley.Lights.Pois$Time=="BEFORE")]           <- Alley.Lights.Pois$Battery.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$CriminalDamage[which(Alley.Lights.Pois$Time=="BEFORE")]    <- Alley.Lights.Pois$CriminalDamage.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$MotorVehicleTheft[which(Alley.Lights.Pois$Time=="BEFORE")] <- Alley.Lights.Pois$MotorVehicleTheft.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$Robbery[which(Alley.Lights.Pois$Time=="BEFORE")]           <- Alley.Lights.Pois$Robbery.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$Assault[which(Alley.Lights.Pois$Time=="BEFORE")]           <- Alley.Lights.Pois$Assault.Before[which(Alley.Lights.Pois$Time=="BEFORE")]
Alley.Lights.Pois$Duration[which(Alley.Lights.Pois$Time=="BEFORE")]          <- rep(30, length(which(Alley.Lights.Pois$Time=="BEFORE")))

Alley.Lights.Pois$AllCrimes[which(Alley.Lights.Pois$Time=="DURING")]         <- Alley.Lights.Pois$Crimes.All.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$Theft[which(Alley.Lights.Pois$Time=="DURING")]             <- Alley.Lights.Pois$Theft.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$Narcotics[which(Alley.Lights.Pois$Time=="DURING")]         <- Alley.Lights.Pois$Narcotics.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$Battery[which(Alley.Lights.Pois$Time=="DURING")]           <- Alley.Lights.Pois$Battery.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$CriminalDamage[which(Alley.Lights.Pois$Time=="DURING")]    <- Alley.Lights.Pois$CriminalDamage.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$MotorVehicleTheft[which(Alley.Lights.Pois$Time=="DURING")] <- Alley.Lights.Pois$MotorVehicleTheft.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$Robbery[which(Alley.Lights.Pois$Time=="DURING")]           <- Alley.Lights.Pois$Robbery.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$Assault[which(Alley.Lights.Pois$Time=="DURING")]           <- Alley.Lights.Pois$Assault.During[which(Alley.Lights.Pois$Time=="DURING")]
Alley.Lights.Pois$Duration[which(Alley.Lights.Pois$Time=="DURING")]          <- Alley.Lights.Pois$OutageDuration[which(Alley.Lights.Pois$Time=="DURING")]

Alley.Lights.Pois$AllCrimes[which(Alley.Lights.Pois$Time=="AFTER")]         <- Alley.Lights.Pois$Crimes.All.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$Theft[which(Alley.Lights.Pois$Time=="AFTER")]             <- Alley.Lights.Pois$Theft.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$Narcotics[which(Alley.Lights.Pois$Time=="AFTER")]         <- Alley.Lights.Pois$Narcotics.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$Battery[which(Alley.Lights.Pois$Time=="AFTER")]           <- Alley.Lights.Pois$Battery.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$CriminalDamage[which(Alley.Lights.Pois$Time=="AFTER")]    <- Alley.Lights.Pois$CriminalDamage.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$MotorVehicleTheft[which(Alley.Lights.Pois$Time=="AFTER")] <- Alley.Lights.Pois$MotorVehicleTheft.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$Robbery[which(Alley.Lights.Pois$Time=="AFTER")]           <- Alley.Lights.Pois$Robbery.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$Assault[which(Alley.Lights.Pois$Time=="AFTER")]           <- Alley.Lights.Pois$Assault.After[which(Alley.Lights.Pois$Time=="AFTER")]
Alley.Lights.Pois$Duration[which(Alley.Lights.Pois$Time=="AFTER")]          <- Alley.Lights.Pois$After.Period.Duration[which(Alley.Lights.Pois$Time=="AFTER")]

Alley.Lights.Pois <- Alley.Lights.Pois[,c("Type", "Service.Request.No","DateCreated", "DateCompleted", "Location", "Outcome", "x_coord", "y_coord", "zip_code", "ward",
                                           "police_district", "community_area", "Time", "AllCrimes", "Theft", "Narcotics", "Battery", "CriminalDamage", 
                                           "MotorVehicleTheft", "Robbery", "Assault", "Duration", "OutageDuration")]
Alley.Lights.Pois$OutageInd <- numeric(nrow(Alley.Lights.Pois))
Alley.Lights.Pois$OutageInd[which(Alley.Lights.Pois$Time=="DURING")] <- rep(1, length(which(Alley.Lights.Pois$Time=="DURING")))


Fit.Alley.AllCrimes         <- glmmboot(AllCrimes         ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Fit.Alley.Theft             <- glmmboot(Theft             ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Fit.Alley.Narcotics         <- glmmboot(Narcotics         ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Fit.Alley.Battery           <- glmmboot(Battery           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Fit.Alley.CriminalDamage    <- glmmboot(CriminalDamage    ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Fit.Alley.MotorVehicleTheft <- glmmboot(MotorVehicleTheft ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Fit.Alley.Robbery           <- glmmboot(Robbery           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Fit.Alley.Assault           <- glmmboot(Assault           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Alley.Lights.Pois)
Summary.Table.Alley$PctDiff <- numeric(nrow(Summary.Table.Alley))
Summary.Table.Alley$LL <- numeric(nrow(Summary.Table.Alley))
Summary.Table.Alley$UL <- numeric(nrow(Summary.Table.Alley))
Summary.Table.Alley$p.value <- numeric(nrow(Summary.Table.Alley))
Summary.Table.Alley$star <- character(nrow(Summary.Table.Alley))
Summary.Table.Alley$PctDiff[1]  <- 100*exp(Fit.Alley.AllCrimes$coef[1])-100
Summary.Table.Alley$PctDiff[2]  <- 100*exp(Fit.Alley.Theft$coef[1])-100
Summary.Table.Alley$PctDiff[3]  <- 100*exp(Fit.Alley.Narcotics$coef[1])-100
Summary.Table.Alley$PctDiff[4]  <- 100*exp(Fit.Alley.Battery$coef[1])-100
Summary.Table.Alley$PctDiff[5]  <- 100*exp(Fit.Alley.CriminalDamage$coef[1])-100
Summary.Table.Alley$PctDiff[6]  <- 100*exp(Fit.Alley.MotorVehicleTheft$coef[1])-100
Summary.Table.Alley$PctDiff[7]  <- 100*exp(Fit.Alley.Robbery$coef[1])-100
Summary.Table.Alley$PctDiff[8]  <- 100*exp(Fit.Alley.Assault$coef[1])-100
Summary.Table.Alley$p.value[1]  <- 2*pt(-abs(Fit.Alley.AllCrimes$coef[1]/Fit.Alley.AllCrimes$sd[1]),df=length(Fit.Alley.AllCrimes$pred-1))
Summary.Table.Alley$p.value[2]  <- 2*pt(-abs(Fit.Alley.Theft$coef[1]/Fit.Alley.Theft$sd[1]),df=length(Fit.Alley.Theft$pred-1))
Summary.Table.Alley$p.value[3]  <- 2*pt(-abs(Fit.Alley.Narcotics$coef[1]/Fit.Alley.Narcotics$sd[1]),df=length(Fit.Alley.Narcotics$pred-1))
Summary.Table.Alley$p.value[4]  <- 2*pt(-abs(Fit.Alley.Battery$coef[1]/Fit.Alley.Battery$sd[1]),df=length(Fit.Alley.Battery$pred-1))
Summary.Table.Alley$p.value[5]  <- 2*pt(-abs(Fit.Alley.CriminalDamage$coef[1]/Fit.Alley.CriminalDamage$sd[1]),df=length(Fit.Alley.CriminalDamage$pred-1))
Summary.Table.Alley$p.value[6]  <- 2*pt(-abs(Fit.Alley.MotorVehicleTheft$coef[1]/Fit.Alley.MotorVehicleTheft$sd[1]),df=length(Fit.Alley.MotorVehicleTheft$pred-1))
Summary.Table.Alley$p.value[7]  <- 2*pt(-abs(Fit.Alley.Robbery$coef[1]/Fit.Alley.Robbery$sd[1]),df=length(Fit.Alley.Robbery$pred-1))
Summary.Table.Alley$p.value[8]  <- 2*pt(-abs(Fit.Alley.Assault$coef[1]/Fit.Alley.Assault$sd[1]),df=length(Fit.Alley.Assault$pred-1))

Summary.Table.Alley$LL[1]  <- 100*exp(Fit.Alley.AllCrimes$coef[1]         - qt(.975, length(Fit.Alley.AllCrimes$pred-1))        *Fit.Alley.AllCrimes$sd[1])        -100
Summary.Table.Alley$LL[2]  <- 100*exp(Fit.Alley.Theft$coef[1]             - qt(.975, length(Fit.Alley.Theft$pred-1))           *Fit.Alley.Theft$sd[1])           -100
Summary.Table.Alley$LL[3]  <- 100*exp(Fit.Alley.Narcotics$coef[1]         - qt(.975, length(Fit.Alley.Narcotics$pred-1))        *Fit.Alley.Narcotics$sd[1])        -100
Summary.Table.Alley$LL[4]  <- 100*exp(Fit.Alley.Battery$coef[1]           - qt(.975, length(Fit.Alley.Battery$pred-1))          *Fit.Alley.Battery$sd[1])          -100
Summary.Table.Alley$LL[5]  <- 100*exp(Fit.Alley.CriminalDamage$coef[1]    - qt(.975, length(Fit.Alley.CriminalDamage$pred-1))   *Fit.Alley.CriminalDamage$sd[1])   -100
Summary.Table.Alley$LL[6]  <- 100*exp(Fit.Alley.MotorVehicleTheft$coef[1] - qt(.975, length(Fit.Alley.MotorVehicleTheft$pred-1))*Fit.Alley.MotorVehicleTheft$sd[1])-100
Summary.Table.Alley$LL[7]  <- 100*exp(Fit.Alley.Robbery$coef[1]           - qt(.975, length(Fit.Alley.Robbery$pred-1))          *Fit.Alley.Robbery$sd[1])          -100
Summary.Table.Alley$LL[8]  <- 100*exp(Fit.Alley.Assault$coef[1]           - qt(.975, length(Fit.Alley.Assault$pred-1))          *Fit.Alley.Assault$sd[1])          -100
Summary.Table.Alley$UL[1]  <- 100*exp(Fit.Alley.AllCrimes$coef[1]         + qt(.975, length(Fit.Alley.AllCrimes$pred-1))        *Fit.Alley.AllCrimes$sd[1])        -100
Summary.Table.Alley$UL[2]  <- 100*exp(Fit.Alley.Theft$coef[1]             + qt(.975, length(Fit.Alley.Theft$pred-1))           *Fit.Alley.Theft$sd[1])           -100
Summary.Table.Alley$UL[3]  <- 100*exp(Fit.Alley.Narcotics$coef[1]         + qt(.975, length(Fit.Alley.Narcotics$pred-1))        *Fit.Alley.Narcotics$sd[1])        -100
Summary.Table.Alley$UL[4]  <- 100*exp(Fit.Alley.Battery$coef[1]           + qt(.975, length(Fit.Alley.Battery$pred-1))          *Fit.Alley.Battery$sd[1])          -100
Summary.Table.Alley$UL[5]  <- 100*exp(Fit.Alley.CriminalDamage$coef[1]    + qt(.975, length(Fit.Alley.CriminalDamage$pred-1))   *Fit.Alley.CriminalDamage$sd[1])   -100
Summary.Table.Alley$UL[6]  <- 100*exp(Fit.Alley.MotorVehicleTheft$coef[1] + qt(.975, length(Fit.Alley.MotorVehicleTheft$pred-1))*Fit.Alley.MotorVehicleTheft$sd[1])-100
Summary.Table.Alley$UL[7]  <- 100*exp(Fit.Alley.Robbery$coef[1]           + qt(.975, length(Fit.Alley.Robbery$pred-1))          *Fit.Alley.Robbery$sd[1])          -100
Summary.Table.Alley$UL[8]  <- 100*exp(Fit.Alley.Assault$coef[1]           + qt(.975, length(Fit.Alley.Assault$pred-1))          *Fit.Alley.Assault$sd[1])          -100
Summary.Table.Alley$star[which(Summary.Table.Alley$p.value<0.01)]                                     <- rep("**", length(which(Summary.Table.Alley$p.value<0.01)))
Summary.Table.Alley$star[which(Summary.Table.Alley$p.value<0.05 & Summary.Table.Alley$p.value>=0.01)] <- rep("*" , length(which(Summary.Table.Alley$p.value<0.05 & Summary.Table.Alley$p.value>=0.01)))




Street.Lights.OneOut.Pois <- Street.Lights.OneOut[rep(seq(nrow(Street.Lights.OneOut)),3),]
Street.Lights.OneOut.Pois <- Street.Lights.OneOut.Pois[order(Street.Lights.OneOut.Pois$Service.Request.No),]
Street.Lights.OneOut.Pois$Time              <- rep(c("BEFORE", "DURING", "AFTER"), nrow(Street.Lights.OneOut))
Street.Lights.OneOut.Pois$AllCrimes         <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Theft             <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Narcotics         <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Battery           <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$CriminalDamage    <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$MotorVehicleTheft <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Robbery           <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Assault           <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Burglary          <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Homicide          <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$DeceptivePractice <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$Duration          <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$AllCrimes[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]         <- Street.Lights.OneOut.Pois$Crimes.All.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Theft[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]            <- Street.Lights.OneOut.Pois$Theft.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Narcotics[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]         <- Street.Lights.OneOut.Pois$Narcotics.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Battery[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]           <- Street.Lights.OneOut.Pois$Battery.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$CriminalDamage[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]    <- Street.Lights.OneOut.Pois$CriminalDamage.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$MotorVehicleTheft[which(Street.Lights.OneOut.Pois$Time=="BEFORE")] <- Street.Lights.OneOut.Pois$MotorVehicleTheft.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Robbery[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]           <- Street.Lights.OneOut.Pois$Robbery.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Assault[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]           <- Street.Lights.OneOut.Pois$Assault.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Burglary[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]          <- Street.Lights.OneOut.Pois$Burglary.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Homicide[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]          <- Street.Lights.OneOut.Pois$Homicide.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$DeceptivePractice[which(Street.Lights.OneOut.Pois$Time=="BEFORE")] <- Street.Lights.OneOut.Pois$DeceptivePractice.Before[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]
Street.Lights.OneOut.Pois$Duration[which(Street.Lights.OneOut.Pois$Time=="BEFORE")]          <- rep(30, length(which(Street.Lights.OneOut.Pois$Time=="BEFORE")))

Street.Lights.OneOut.Pois$AllCrimes[which(Street.Lights.OneOut.Pois$Time=="DURING")]         <- Street.Lights.OneOut.Pois$Crimes.All.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Theft[which(Street.Lights.OneOut.Pois$Time=="DURING")]            <- Street.Lights.OneOut.Pois$Theft.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Narcotics[which(Street.Lights.OneOut.Pois$Time=="DURING")]         <- Street.Lights.OneOut.Pois$Narcotics.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Battery[which(Street.Lights.OneOut.Pois$Time=="DURING")]           <- Street.Lights.OneOut.Pois$Battery.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$CriminalDamage[which(Street.Lights.OneOut.Pois$Time=="DURING")]    <- Street.Lights.OneOut.Pois$CriminalDamage.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$MotorVehicleTheft[which(Street.Lights.OneOut.Pois$Time=="DURING")] <- Street.Lights.OneOut.Pois$MotorVehicleTheft.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Robbery[which(Street.Lights.OneOut.Pois$Time=="DURING")]           <- Street.Lights.OneOut.Pois$Robbery.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Assault[which(Street.Lights.OneOut.Pois$Time=="DURING")]           <- Street.Lights.OneOut.Pois$Assault.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Burglary[which(Street.Lights.OneOut.Pois$Time=="DURING")]          <- Street.Lights.OneOut.Pois$Burglary.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Homicide[which(Street.Lights.OneOut.Pois$Time=="DURING")]          <- Street.Lights.OneOut.Pois$Homicide.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$DeceptivePractice[which(Street.Lights.OneOut.Pois$Time=="DURING")] <- Street.Lights.OneOut.Pois$DeceptivePractice.During[which(Street.Lights.OneOut.Pois$Time=="DURING")]
Street.Lights.OneOut.Pois$Duration[which(Street.Lights.OneOut.Pois$Time=="DURING")]          <- Street.Lights.OneOut.Pois$OutageDuration[which(Street.Lights.OneOut.Pois$Time=="DURING")]

Street.Lights.OneOut.Pois$AllCrimes[which(Street.Lights.OneOut.Pois$Time=="AFTER")]         <- Street.Lights.OneOut.Pois$Crimes.All.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Theft[which(Street.Lights.OneOut.Pois$Time=="AFTER")]            <- Street.Lights.OneOut.Pois$Theft.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Narcotics[which(Street.Lights.OneOut.Pois$Time=="AFTER")]         <- Street.Lights.OneOut.Pois$Narcotics.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Battery[which(Street.Lights.OneOut.Pois$Time=="AFTER")]           <- Street.Lights.OneOut.Pois$Battery.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$CriminalDamage[which(Street.Lights.OneOut.Pois$Time=="AFTER")]    <- Street.Lights.OneOut.Pois$CriminalDamage.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$MotorVehicleTheft[which(Street.Lights.OneOut.Pois$Time=="AFTER")] <- Street.Lights.OneOut.Pois$MotorVehicleTheft.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Robbery[which(Street.Lights.OneOut.Pois$Time=="AFTER")]           <- Street.Lights.OneOut.Pois$Robbery.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Assault[which(Street.Lights.OneOut.Pois$Time=="AFTER")]           <- Street.Lights.OneOut.Pois$Assault.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Burglary[which(Street.Lights.OneOut.Pois$Time=="AFTER")]          <- Street.Lights.OneOut.Pois$Burglary.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Homicide[which(Street.Lights.OneOut.Pois$Time=="AFTER")]          <- Street.Lights.OneOut.Pois$Homicide.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$DeceptivePractice[which(Street.Lights.OneOut.Pois$Time=="AFTER")] <- Street.Lights.OneOut.Pois$DeceptivePractice.After[which(Street.Lights.OneOut.Pois$Time=="AFTER")]
Street.Lights.OneOut.Pois$Duration[which(Street.Lights.OneOut.Pois$Time=="AFTER")]          <- Street.Lights.OneOut.Pois$After.Period.Duration[which(Street.Lights.OneOut.Pois$Time=="AFTER")]

Street.Lights.OneOut.Pois <- Street.Lights.OneOut.Pois[,c("Type", "Service.Request.No","DateCreated", "DateCompleted", "Location", "Outcome", "x_coord", "y_coord", "zip_code", "ward",
                                          "police_district", "community_area", "Time", "AllCrimes", "Theft", "Narcotics", "Battery", "CriminalDamage", 
                                          "MotorVehicleTheft", "Robbery", "Assault", "Burglary", "Homicide", "DeceptivePractice", "Duration", "OutageDuration")]
Street.Lights.OneOut.Pois$OutageInd <- numeric(nrow(Street.Lights.OneOut.Pois))
Street.Lights.OneOut.Pois$OutageInd[which(Street.Lights.OneOut.Pois$Time=="DURING")] <- rep(1, length(which(Street.Lights.OneOut.Pois$Time=="DURING")))



#Poisson glm Models to Estimate St. Error of % Diff's
Fit.OneOut.AllCrimes         <- glmmboot(AllCrimes         ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.Theft             <- glmmboot(Theft             ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.Narcotics         <- glmmboot(Narcotics         ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.Battery           <- glmmboot(Battery           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.CriminalDamage    <- glmmboot(CriminalDamage    ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.MotorVehicleTheft <- glmmboot(MotorVehicleTheft ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.Robbery           <- glmmboot(Robbery           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.Assault           <- glmmboot(Assault           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.Burglary          <- glmmboot(Burglary          ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.Homicide          <- glmmboot(Homicide          ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Fit.OneOut.DeceptivePractice <- glmmboot(DeceptivePractice ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.OneOut.Pois)
Summary.Table.OneOut$PctDiff <- numeric(nrow(Summary.Table.OneOut))
Summary.Table.OneOut$LL <- numeric(nrow(Summary.Table.OneOut))
Summary.Table.OneOut$UL <- numeric(nrow(Summary.Table.OneOut))
Summary.Table.OneOut$p.value <- numeric(nrow(Summary.Table.OneOut))
Summary.Table.OneOut$star <- character(nrow(Summary.Table.OneOut))
Summary.Table.OneOut$PctDiff[1]  <- 100*exp(Fit.OneOut.AllCrimes$coef[1])-100
Summary.Table.OneOut$PctDiff[2]  <- 100*exp(Fit.OneOut.Theft$coef[1])-100
Summary.Table.OneOut$PctDiff[3]  <- 100*exp(Fit.OneOut.Narcotics$coef[1])-100
Summary.Table.OneOut$PctDiff[4]  <- 100*exp(Fit.OneOut.Battery$coef[1])-100
Summary.Table.OneOut$PctDiff[5]  <- 100*exp(Fit.OneOut.CriminalDamage$coef[1])-100
Summary.Table.OneOut$PctDiff[6]  <- 100*exp(Fit.OneOut.MotorVehicleTheft$coef[1])-100
Summary.Table.OneOut$PctDiff[7]  <- 100*exp(Fit.OneOut.Robbery$coef[1])-100
Summary.Table.OneOut$PctDiff[8]  <- 100*exp(Fit.OneOut.Assault$coef[1])-100
Summary.Table.OneOut$PctDiff[9]  <- 100*exp(Fit.OneOut.Burglary$coef[1])-100
Summary.Table.OneOut$PctDiff[10] <- 100*exp(Fit.OneOut.Homicide$coef[1])-100
Summary.Table.OneOut$PctDiff[11] <- 100*exp(Fit.OneOut.DeceptivePractice$coef[1])-100
Summary.Table.OneOut$p.value[1]  <- 2*pt(-abs(Fit.OneOut.AllCrimes$coef[1]/Fit.OneOut.AllCrimes$sd[1]),df=length(Fit.OneOut.AllCrimes$pred-1))
Summary.Table.OneOut$p.value[2]  <- 2*pt(-abs(Fit.OneOut.Theft$coef[1]/Fit.OneOut.Theft$sd[1]),df=length(Fit.OneOut.Theft$pred-1))
Summary.Table.OneOut$p.value[3]  <- 2*pt(-abs(Fit.OneOut.Narcotics$coef[1]/Fit.OneOut.Narcotics$sd[1]),df=length(Fit.OneOut.Narcotics$pred-1))
Summary.Table.OneOut$p.value[4]  <- 2*pt(-abs(Fit.OneOut.Battery$coef[1]/Fit.OneOut.Battery$sd[1]),df=length(Fit.OneOut.Battery$pred-1))
Summary.Table.OneOut$p.value[5]  <- 2*pt(-abs(Fit.OneOut.CriminalDamage$coef[1]/Fit.OneOut.CriminalDamage$sd[1]),df=length(Fit.OneOut.CriminalDamage$pred-1))
Summary.Table.OneOut$p.value[6]  <- 2*pt(-abs(Fit.OneOut.MotorVehicleTheft$coef[1]/Fit.OneOut.MotorVehicleTheft$sd[1]),df=length(Fit.OneOut.MotorVehicleTheft$pred-1))
Summary.Table.OneOut$p.value[7]  <- 2*pt(-abs(Fit.OneOut.Robbery$coef[1]/Fit.OneOut.Robbery$sd[1]),df=length(Fit.OneOut.Robbery$pred-1))
Summary.Table.OneOut$p.value[8]  <- 2*pt(-abs(Fit.OneOut.Assault$coef[1]/Fit.OneOut.Assault$sd[1]),df=length(Fit.OneOut.Assault$pred-1))
Summary.Table.OneOut$p.value[9]  <- 2*pt(-abs(Fit.OneOut.Burglary$coef[1]/Fit.OneOut.Burglary$sd[1]),df=length(Fit.OneOut.Burglary$pred-1))
Summary.Table.OneOut$p.value[10] <- 2*pt(-abs(Fit.OneOut.Homicide$coef[1]/Fit.OneOut.Homicide$sd[1]),df=length(Fit.OneOut.Homicide$pred-1))
Summary.Table.OneOut$p.value[11] <- 2*pt(-abs(Fit.OneOut.DeceptivePractice$coef[1]/Fit.OneOut.DeceptivePractice$sd[1]),df=length(Fit.OneOut.DeceptivePractice$pred-1))
Summary.Table.OneOut$LL[1]  <- 100*exp(Fit.OneOut.AllCrimes$coef[1]         - qt(.975, length(Fit.OneOut.AllCrimes$pred-1))        *Fit.OneOut.AllCrimes$sd[1])        -100
Summary.Table.OneOut$LL[2]  <- 100*exp(Fit.OneOut.Theft$coef[1]            - qt(.975, length(Fit.OneOut.Theft$pred-1))           *Fit.OneOut.Theft$sd[1])           -100
Summary.Table.OneOut$LL[3]  <- 100*exp(Fit.OneOut.Narcotics$coef[1]         - qt(.975, length(Fit.OneOut.Narcotics$pred-1))        *Fit.OneOut.Narcotics$sd[1])        -100
Summary.Table.OneOut$LL[4]  <- 100*exp(Fit.OneOut.Battery$coef[1]           - qt(.975, length(Fit.OneOut.Battery$pred-1))          *Fit.OneOut.Battery$sd[1])          -100
Summary.Table.OneOut$LL[5]  <- 100*exp(Fit.OneOut.CriminalDamage$coef[1]    - qt(.975, length(Fit.OneOut.CriminalDamage$pred-1))   *Fit.OneOut.CriminalDamage$sd[1])   -100
Summary.Table.OneOut$LL[6]  <- 100*exp(Fit.OneOut.MotorVehicleTheft$coef[1] - qt(.975, length(Fit.OneOut.MotorVehicleTheft$pred-1))*Fit.OneOut.MotorVehicleTheft$sd[1])-100
Summary.Table.OneOut$LL[7]  <- 100*exp(Fit.OneOut.Robbery$coef[1]           - qt(.975, length(Fit.OneOut.Robbery$pred-1))          *Fit.OneOut.Robbery$sd[1])          -100
Summary.Table.OneOut$LL[8]  <- 100*exp(Fit.OneOut.Assault$coef[1]           - qt(.975, length(Fit.OneOut.Assault$pred-1))          *Fit.OneOut.Assault$sd[1])          -100
Summary.Table.OneOut$LL[9]  <- 100*exp(Fit.OneOut.Burglary$coef[1]          - qt(.975, length(Fit.OneOut.Burglary$pred-1))         *Fit.OneOut.Burglary$sd[1])         -100
Summary.Table.OneOut$LL[10] <- 100*exp(Fit.OneOut.Homicide$coef[1]          - qt(.975, length(Fit.OneOut.Homicide$pred-1))         *Fit.OneOut.Homicide$sd[1])         -100
Summary.Table.OneOut$LL[11] <- 100*exp(Fit.OneOut.DeceptivePractice$coef[1] - qt(.975, length(Fit.OneOut.DeceptivePractice$pred-1))*Fit.OneOut.DeceptivePractice$sd[1])-100
Summary.Table.OneOut$UL[1]  <- 100*exp(Fit.OneOut.AllCrimes$coef[1]         + qt(.975, length(Fit.OneOut.AllCrimes$pred-1))        *Fit.OneOut.AllCrimes$sd[1])        -100
Summary.Table.OneOut$UL[2]  <- 100*exp(Fit.OneOut.Theft$coef[1]             + qt(.975, length(Fit.OneOut.Theft$pred-1))           *Fit.OneOut.Theft$sd[1])           -100
Summary.Table.OneOut$UL[3]  <- 100*exp(Fit.OneOut.Narcotics$coef[1]         + qt(.975, length(Fit.OneOut.Narcotics$pred-1))        *Fit.OneOut.Narcotics$sd[1])        -100
Summary.Table.OneOut$UL[4]  <- 100*exp(Fit.OneOut.Battery$coef[1]           + qt(.975, length(Fit.OneOut.Battery$pred-1))          *Fit.OneOut.Battery$sd[1])          -100
Summary.Table.OneOut$UL[5]  <- 100*exp(Fit.OneOut.CriminalDamage$coef[1]    + qt(.975, length(Fit.OneOut.CriminalDamage$pred-1))   *Fit.OneOut.CriminalDamage$sd[1])   -100
Summary.Table.OneOut$UL[6]  <- 100*exp(Fit.OneOut.MotorVehicleTheft$coef[1] + qt(.975, length(Fit.OneOut.MotorVehicleTheft$pred-1))*Fit.OneOut.MotorVehicleTheft$sd[1])-100
Summary.Table.OneOut$UL[7]  <- 100*exp(Fit.OneOut.Robbery$coef[1]           + qt(.975, length(Fit.OneOut.Robbery$pred-1))          *Fit.OneOut.Robbery$sd[1])          -100
Summary.Table.OneOut$UL[8]  <- 100*exp(Fit.OneOut.Assault$coef[1]           + qt(.975, length(Fit.OneOut.Assault$pred-1))          *Fit.OneOut.Assault$sd[1])          -100
Summary.Table.OneOut$UL[9]  <- 100*exp(Fit.OneOut.Burglary$coef[1]          + qt(.975, length(Fit.OneOut.Burglary$pred-1))         *Fit.OneOut.Burglary$sd[1])         -100
Summary.Table.OneOut$UL[10] <- 100*exp(Fit.OneOut.Homicide$coef[1]          + qt(.975, length(Fit.OneOut.Homicide$pred-1))         *Fit.OneOut.Homicide$sd[1])         -100
Summary.Table.OneOut$UL[11] <- 100*exp(Fit.OneOut.DeceptivePractice$coef[1] + qt(.975, length(Fit.OneOut.DeceptivePractice$pred-1))*Fit.OneOut.DeceptivePractice$sd[1])-100
Summary.Table.OneOut$star[which(Summary.Table.OneOut$p.value<0.01)]                                     <- rep("**", length(which(Summary.Table.OneOut$p.value<0.01)))
Summary.Table.OneOut$star[which(Summary.Table.OneOut$p.value<0.05 & Summary.Table.OneOut$p.value>=0.01)] <- rep("*" , length(which(Summary.Table.OneOut$p.value<0.05 & Summary.Table.OneOut$p.value>=0.01)))




Street.Lights.AllOut.Pois <- Street.Lights.AllOut[rep(seq(nrow(Street.Lights.AllOut)),3),]
Street.Lights.AllOut.Pois <- Street.Lights.AllOut.Pois[order(Street.Lights.AllOut.Pois$Service.Request.No),]
Street.Lights.AllOut.Pois$Time              <- rep(c("BEFORE", "DURING", "AFTER"), nrow(Street.Lights.AllOut))
Street.Lights.AllOut.Pois$AllCrimes         <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Theft            <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Narcotics         <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Battery           <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$CriminalDamage    <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$MotorVehicleTheft <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Robbery           <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Assault           <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Burglary          <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Homicide          <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$DeceptivePractice <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$Duration          <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$AllCrimes[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]         <- Street.Lights.AllOut.Pois$Crimes.All.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Theft[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]             <- Street.Lights.AllOut.Pois$Theft.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Narcotics[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]         <- Street.Lights.AllOut.Pois$Narcotics.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Battery[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]           <- Street.Lights.AllOut.Pois$Battery.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$CriminalDamage[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]    <- Street.Lights.AllOut.Pois$CriminalDamage.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$MotorVehicleTheft[which(Street.Lights.AllOut.Pois$Time=="BEFORE")] <- Street.Lights.AllOut.Pois$MotorVehicleTheft.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Robbery[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]           <- Street.Lights.AllOut.Pois$Robbery.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Assault[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]           <- Street.Lights.AllOut.Pois$Assault.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Burglary[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]          <- Street.Lights.AllOut.Pois$Burglary.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Homicide[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]          <- Street.Lights.AllOut.Pois$Homicide.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$DeceptivePractice[which(Street.Lights.AllOut.Pois$Time=="BEFORE")] <- Street.Lights.AllOut.Pois$DeceptivePractice.Before[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]
Street.Lights.AllOut.Pois$Duration[which(Street.Lights.AllOut.Pois$Time=="BEFORE")]          <- rep(30, length(which(Street.Lights.AllOut.Pois$Time=="BEFORE")))

Street.Lights.AllOut.Pois$AllCrimes[which(Street.Lights.AllOut.Pois$Time=="DURING")]         <- Street.Lights.AllOut.Pois$Crimes.All.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Theft[which(Street.Lights.AllOut.Pois$Time=="DURING")]             <- Street.Lights.AllOut.Pois$Theft.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Narcotics[which(Street.Lights.AllOut.Pois$Time=="DURING")]         <- Street.Lights.AllOut.Pois$Narcotics.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Battery[which(Street.Lights.AllOut.Pois$Time=="DURING")]           <- Street.Lights.AllOut.Pois$Battery.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$CriminalDamage[which(Street.Lights.AllOut.Pois$Time=="DURING")]    <- Street.Lights.AllOut.Pois$CriminalDamage.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$MotorVehicleTheft[which(Street.Lights.AllOut.Pois$Time=="DURING")] <- Street.Lights.AllOut.Pois$MotorVehicleTheft.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Robbery[which(Street.Lights.AllOut.Pois$Time=="DURING")]           <- Street.Lights.AllOut.Pois$Robbery.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Assault[which(Street.Lights.AllOut.Pois$Time=="DURING")]           <- Street.Lights.AllOut.Pois$Assault.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Burglary[which(Street.Lights.AllOut.Pois$Time=="DURING")]          <- Street.Lights.AllOut.Pois$Burglary.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Homicide[which(Street.Lights.AllOut.Pois$Time=="DURING")]          <- Street.Lights.AllOut.Pois$Homicide.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$DeceptivePractice[which(Street.Lights.AllOut.Pois$Time=="DURING")] <- Street.Lights.AllOut.Pois$DeceptivePractice.During[which(Street.Lights.AllOut.Pois$Time=="DURING")]
Street.Lights.AllOut.Pois$Duration[which(Street.Lights.AllOut.Pois$Time=="DURING")]          <- Street.Lights.AllOut.Pois$OutageDuration[which(Street.Lights.AllOut.Pois$Time=="DURING")]

Street.Lights.AllOut.Pois$AllCrimes[which(Street.Lights.AllOut.Pois$Time=="AFTER")]         <- Street.Lights.AllOut.Pois$Crimes.All.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Theft[which(Street.Lights.AllOut.Pois$Time=="AFTER")]             <- Street.Lights.AllOut.Pois$Theft.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Narcotics[which(Street.Lights.AllOut.Pois$Time=="AFTER")]         <- Street.Lights.AllOut.Pois$Narcotics.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Battery[which(Street.Lights.AllOut.Pois$Time=="AFTER")]           <- Street.Lights.AllOut.Pois$Battery.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$CriminalDamage[which(Street.Lights.AllOut.Pois$Time=="AFTER")]    <- Street.Lights.AllOut.Pois$CriminalDamage.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$MotorVehicleTheft[which(Street.Lights.AllOut.Pois$Time=="AFTER")] <- Street.Lights.AllOut.Pois$MotorVehicleTheft.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Robbery[which(Street.Lights.AllOut.Pois$Time=="AFTER")]           <- Street.Lights.AllOut.Pois$Robbery.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Assault[which(Street.Lights.AllOut.Pois$Time=="AFTER")]           <- Street.Lights.AllOut.Pois$Assault.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Burglary[which(Street.Lights.AllOut.Pois$Time=="AFTER")]          <- Street.Lights.AllOut.Pois$Burglary.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Homicide[which(Street.Lights.AllOut.Pois$Time=="AFTER")]          <- Street.Lights.AllOut.Pois$Homicide.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$DeceptivePractice[which(Street.Lights.AllOut.Pois$Time=="AFTER")] <- Street.Lights.AllOut.Pois$DeceptivePractice.After[which(Street.Lights.AllOut.Pois$Time=="AFTER")]
Street.Lights.AllOut.Pois$Duration[which(Street.Lights.AllOut.Pois$Time=="AFTER")]          <- Street.Lights.AllOut.Pois$After.Period.Duration[which(Street.Lights.AllOut.Pois$Time=="AFTER")]

Street.Lights.AllOut.Pois <- Street.Lights.AllOut.Pois[,c("Type", "Service.Request.No","DateCreated", "DateCompleted", "Location", "Outcome", "x_coord", "y_coord", "zip_code", "ward",
                                                          "police_district", "community_area", "Time", "AllCrimes", "Theft", "Narcotics", "Battery", "CriminalDamage", 
                                                          "MotorVehicleTheft", "Robbery", "Assault", "Burglary", "Homicide", "DeceptivePractice", "Duration", "OutageDuration")]
Street.Lights.AllOut.Pois$OutageInd <- numeric(nrow(Street.Lights.AllOut.Pois))
Street.Lights.AllOut.Pois$OutageInd[which(Street.Lights.AllOut.Pois$Time=="DURING")] <- rep(1, length(which(Street.Lights.AllOut.Pois$Time=="DURING")))

Fit.AllOut.AllCrimes         <- glmmboot(AllCrimes         ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.Theft             <- glmmboot(Theft             ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.Narcotics         <- glmmboot(Narcotics         ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.Battery           <- glmmboot(Battery           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.CriminalDamage    <- glmmboot(CriminalDamage    ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.MotorVehicleTheft <- glmmboot(MotorVehicleTheft ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.Robbery           <- glmmboot(Robbery           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.Assault           <- glmmboot(Assault           ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.Burglary          <- glmmboot(Burglary          ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.Homicide          <- glmmboot(Homicide          ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Fit.AllOut.DeceptivePractice <- glmmboot(DeceptivePractice ~ offset(log(Duration)) + OutageInd, cluster=Service.Request.No, family=poisson(link=log), data=Street.Lights.AllOut.Pois)
Summary.Table.AllOut$PctDiff <- numeric(nrow(Summary.Table.AllOut))
Summary.Table.AllOut$LL <- numeric(nrow(Summary.Table.AllOut))
Summary.Table.AllOut$UL <- numeric(nrow(Summary.Table.AllOut))
Summary.Table.AllOut$p.value <- numeric(nrow(Summary.Table.AllOut))
Summary.Table.AllOut$star <- character(nrow(Summary.Table.AllOut))
Summary.Table.AllOut$PctDiff[1]  <- 100*exp(Fit.AllOut.AllCrimes$coef[1])-100
Summary.Table.AllOut$PctDiff[2]  <- 100*exp(Fit.AllOut.Theft$coef[1])-100
Summary.Table.AllOut$PctDiff[3]  <- 100*exp(Fit.AllOut.Narcotics$coef[1])-100
Summary.Table.AllOut$PctDiff[4]  <- 100*exp(Fit.AllOut.Battery$coef[1])-100
Summary.Table.AllOut$PctDiff[5]  <- 100*exp(Fit.AllOut.CriminalDamage$coef[1])-100
Summary.Table.AllOut$PctDiff[6]  <- 100*exp(Fit.AllOut.MotorVehicleTheft$coef[1])-100
Summary.Table.AllOut$PctDiff[7]  <- 100*exp(Fit.AllOut.Robbery$coef[1])-100
Summary.Table.AllOut$PctDiff[8]  <- 100*exp(Fit.AllOut.Assault$coef[1])-100
Summary.Table.AllOut$PctDiff[9]  <- 100*exp(Fit.AllOut.Burglary$coef[1])-100
Summary.Table.AllOut$PctDiff[10] <- 100*exp(Fit.AllOut.Homicide$coef[1])-100
Summary.Table.AllOut$PctDiff[11] <- 100*exp(Fit.AllOut.DeceptivePractice$coef[1])-100
Summary.Table.AllOut$p.value[1]  <- 2*pt(-abs(Fit.AllOut.AllCrimes$coef[1]/Fit.AllOut.AllCrimes$sd[1]),df=length(Fit.AllOut.AllCrimes$pred-1))
Summary.Table.AllOut$p.value[2]  <- 2*pt(-abs(Fit.AllOut.Theft$coef[1]/Fit.AllOut.Theft$sd[1]),df=length(Fit.AllOut.Theft$pred-1))
Summary.Table.AllOut$p.value[3]  <- 2*pt(-abs(Fit.AllOut.Narcotics$coef[1]/Fit.AllOut.Narcotics$sd[1]),df=length(Fit.AllOut.Narcotics$pred-1))
Summary.Table.AllOut$p.value[4]  <- 2*pt(-abs(Fit.AllOut.Battery$coef[1]/Fit.AllOut.Battery$sd[1]),df=length(Fit.AllOut.Battery$pred-1))
Summary.Table.AllOut$p.value[5]  <- 2*pt(-abs(Fit.AllOut.CriminalDamage$coef[1]/Fit.AllOut.CriminalDamage$sd[1]),df=length(Fit.AllOut.CriminalDamage$pred-1))
Summary.Table.AllOut$p.value[6]  <- 2*pt(-abs(Fit.AllOut.MotorVehicleTheft$coef[1]/Fit.AllOut.MotorVehicleTheft$sd[1]),df=length(Fit.AllOut.MotorVehicleTheft$pred-1))
Summary.Table.AllOut$p.value[7]  <- 2*pt(-abs(Fit.AllOut.Robbery$coef[1]/Fit.AllOut.Robbery$sd[1]),df=length(Fit.AllOut.Robbery$pred-1))
Summary.Table.AllOut$p.value[8]  <- 2*pt(-abs(Fit.AllOut.Assault$coef[1]/Fit.AllOut.Assault$sd[1]),df=length(Fit.AllOut.Assault$pred-1))
Summary.Table.AllOut$p.value[9]  <- 2*pt(-abs(Fit.AllOut.Burglary$coef[1]/Fit.AllOut.Burglary$sd[1]),df=length(Fit.AllOut.Burglary$pred-1))
Summary.Table.AllOut$p.value[10] <- 2*pt(-abs(Fit.AllOut.Homicide$coef[1]/Fit.AllOut.Homicide$sd[1]),df=length(Fit.AllOut.Homicide$pred-1))
Summary.Table.AllOut$p.value[11] <- 2*pt(-abs(Fit.AllOut.DeceptivePractice$coef[1]/Fit.AllOut.DeceptivePractice$sd[1]),df=length(Fit.AllOut.DeceptivePractice$pred-1))
Summary.Table.AllOut$LL[1]  <- 100*exp(Fit.AllOut.AllCrimes$coef[1]         - qt(.975, length(Fit.AllOut.AllCrimes$pred-1))        *Fit.AllOut.AllCrimes$sd[1])        -100
Summary.Table.AllOut$LL[2]  <- 100*exp(Fit.AllOut.Theft$coef[1]             - qt(.975, length(Fit.AllOut.Theft$pred-1))           *Fit.AllOut.Theft$sd[1])           -100
Summary.Table.AllOut$LL[3]  <- 100*exp(Fit.AllOut.Narcotics$coef[1]         - qt(.975, length(Fit.AllOut.Narcotics$pred-1))        *Fit.AllOut.Narcotics$sd[1])        -100
Summary.Table.AllOut$LL[4]  <- 100*exp(Fit.AllOut.Battery$coef[1]           - qt(.975, length(Fit.AllOut.Battery$pred-1))          *Fit.AllOut.Battery$sd[1])          -100
Summary.Table.AllOut$LL[5]  <- 100*exp(Fit.AllOut.CriminalDamage$coef[1]    - qt(.975, length(Fit.AllOut.CriminalDamage$pred-1))   *Fit.AllOut.CriminalDamage$sd[1])   -100
Summary.Table.AllOut$LL[6]  <- 100*exp(Fit.AllOut.MotorVehicleTheft$coef[1] - qt(.975, length(Fit.AllOut.MotorVehicleTheft$pred-1))*Fit.AllOut.MotorVehicleTheft$sd[1])-100
Summary.Table.AllOut$LL[7]  <- 100*exp(Fit.AllOut.Robbery$coef[1]           - qt(.975, length(Fit.AllOut.Robbery$pred-1))          *Fit.AllOut.Robbery$sd[1])          -100
Summary.Table.AllOut$LL[8]  <- 100*exp(Fit.AllOut.Assault$coef[1]           - qt(.975, length(Fit.AllOut.Assault$pred-1))          *Fit.AllOut.Assault$sd[1])          -100
Summary.Table.AllOut$LL[9]  <- 100*exp(Fit.AllOut.Burglary$coef[1]          - qt(.975, length(Fit.AllOut.Burglary$pred-1))         *Fit.AllOut.Burglary$sd[1])         -100
Summary.Table.AllOut$LL[10] <- 100*exp(Fit.AllOut.Homicide$coef[1]          - qt(.975, length(Fit.AllOut.Homicide$pred-1))         *Fit.AllOut.Homicide$sd[1])         -100
Summary.Table.AllOut$LL[11] <- 100*exp(Fit.AllOut.DeceptivePractice$coef[1] - qt(.975, length(Fit.AllOut.DeceptivePractice$pred-1))*Fit.AllOut.DeceptivePractice$sd[1])-100
Summary.Table.AllOut$UL[1]  <- 100*exp(Fit.AllOut.AllCrimes$coef[1]         + qt(.975, length(Fit.AllOut.AllCrimes$pred-1))        *Fit.AllOut.AllCrimes$sd[1])        -100
Summary.Table.AllOut$UL[2]  <- 100*exp(Fit.AllOut.Theft$coef[1]             + qt(.975, length(Fit.AllOut.Theft$pred-1))           *Fit.AllOut.Theft$sd[1])           -100
Summary.Table.AllOut$UL[3]  <- 100*exp(Fit.AllOut.Narcotics$coef[1]         + qt(.975, length(Fit.AllOut.Narcotics$pred-1))        *Fit.AllOut.Narcotics$sd[1])        -100
Summary.Table.AllOut$UL[4]  <- 100*exp(Fit.AllOut.Battery$coef[1]           + qt(.975, length(Fit.AllOut.Battery$pred-1))          *Fit.AllOut.Battery$sd[1])          -100
Summary.Table.AllOut$UL[5]  <- 100*exp(Fit.AllOut.CriminalDamage$coef[1]    + qt(.975, length(Fit.AllOut.CriminalDamage$pred-1))   *Fit.AllOut.CriminalDamage$sd[1])   -100
Summary.Table.AllOut$UL[6]  <- 100*exp(Fit.AllOut.MotorVehicleTheft$coef[1] + qt(.975, length(Fit.AllOut.MotorVehicleTheft$pred-1))*Fit.AllOut.MotorVehicleTheft$sd[1])-100
Summary.Table.AllOut$UL[7]  <- 100*exp(Fit.AllOut.Robbery$coef[1]           + qt(.975, length(Fit.AllOut.Robbery$pred-1))          *Fit.AllOut.Robbery$sd[1])          -100
Summary.Table.AllOut$UL[8]  <- 100*exp(Fit.AllOut.Assault$coef[1]           + qt(.975, length(Fit.AllOut.Assault$pred-1))          *Fit.AllOut.Assault$sd[1])          -100
Summary.Table.AllOut$UL[9]  <- 100*exp(Fit.AllOut.Burglary$coef[1]          + qt(.975, length(Fit.AllOut.Burglary$pred-1))         *Fit.AllOut.Burglary$sd[1])         -100
Summary.Table.AllOut$UL[10] <- 100*exp(Fit.AllOut.Homicide$coef[1]          + qt(.975, length(Fit.AllOut.Homicide$pred-1))         *Fit.AllOut.Homicide$sd[1])         -100
Summary.Table.AllOut$UL[11] <- 100*exp(Fit.AllOut.DeceptivePractice$coef[1] + qt(.975, length(Fit.AllOut.DeceptivePractice$pred-1))*Fit.AllOut.DeceptivePractice$sd[1])-100
Summary.Table.AllOut$star[which(Summary.Table.AllOut$p.value<0.01)]                                     <- rep("**", length(which(Summary.Table.AllOut$p.value<0.01)))
Summary.Table.AllOut$star[which(Summary.Table.AllOut$p.value<0.05 & Summary.Table.AllOut$p.value>=0.01)] <- rep("*" , length(which(Summary.Table.AllOut$p.value<0.05 & Summary.Table.AllOut$p.value>=0.01)))

Summary.Table.Alley[,-c(10)]  <- round(Summary.Table.Alley[,-c(10)] , digits=3) 
Summary.Table.OneOut[,-c(10)] <- round(Summary.Table.OneOut[,-c(10)], digits=3)
Summary.Table.AllOut[,-c(10)] <- round(Summary.Table.AllOut[,-c(10)], digits=3)

# Export Tables for Community Area Level Estimates
write.csv(Alley.Lights.Pois, "Alley_Lights_and_Crime_for_Comm_Area_Est.csv")
write.csv(Street.Lights.OneOut.Pois, "Street_Lights_One_Out_and_Crime_for_Comm_Area_Est.csv")
write.csv(Street.Lights.AllOut.Pois, "Street_Lights_All_Out_and_Crime_for_Comm_Area_Est.csv")
